(BSR)[PRO] chore: move route_map to routes/
